feat(nvim): replace YCM with deoplete
This replaces YCM autocomplete(+supertab) with deoplete. Since deoplete
relies on Python3, this commit alos adds a packages3.txt for pip3.

Another thing that was necessary is to rebind tab-complete in deoplete
with something else since ultisnips needs tab for expanding and both of
them don't work that well together.

Using deoplete in supertab gets rid of the speed advantages.